--- a/uaclient/tests/test_cli_auto_attach.py
+++ b/uaclient/tests/test_cli_auto_attach.py
@@ -274,14 +274,16 @@ class TestActionAutoAttach:
 
 class TestParser:
     @mock.patch(M_PATH + "contract.get_available_resources")
-    def test_auto_attach_parser_updates_parser_config(self, _m_resources):
+    def test_auto_attach_parser_updates_parser_config(
+        self, _m_resources, FakeConfig
+    ):
         """Update the parser configuration for 'auto-attach'."""
         m_parser = auto_attach_parser(mock.Mock())
         assert "ua auto-attach [flags]" == m_parser.usage
         assert "auto-attach" == m_parser.prog
         assert "Flags" == m_parser._optionals.title
 
-        full_parser = get_parser()
+        full_parser = get_parser(FakeConfig())
         with mock.patch("sys.argv", ["ua", "auto-attach"]):
             args = full_parser.parse_args()
         assert "auto-attach" == args.command