… uaclient/tests/test_cli.py
… uaclient/tests/test_cli_attach.py
… uaclient/tests/test_cli_detach.py
… uaclient/tests/test_cli_disable.py
… uaclient/tests/test_cli_enable.py
… uaclient/tests/test_cli_security_status.py
… uaclient/tests/test_cli_status.py
… uaclient/tests/test_config.py
… uaclient/tests/test_contract.py
… uaclient/tests/test_update_messaging.py